# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for University Student Admission System",
        "title": "DBMS Project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/applications/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a program application for the authenticated student. Checks for missing required profile fields before creation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Apply to an academic program",
                "parameters": [
                    {
                        "description": "Program Application Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_rest.ApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Application submitted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or incomplete profile with missing fields",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to submit application or error fetching requirements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/programs/eligible": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Evaluates authenticated student's academic performance (GPA, HSC group, subject marks) against program admission requirements.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "List eligible programs for authenticated student",
                "responses": {
                    "200": {
                        "description": "List of eligible programs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler_rest.EligibleProgram"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch student academics, subject marks, or programs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_rest.ApplyRequest": {
            "type": "object",
            "properties": {
                "program_id": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_rest.EligibleProgram": {
            "type": "object",
            "properties": {
                "program_id": {
                    "type": "integer"
                },
                "program_name": {
                    "type": "string"
                },
                "university_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}